remove old grammar, rename new grammar
