+++ vllm/grammar.py
@@ -1,171 +1,338 @@
 import collections
-from parsimonious.grammar import Grammar
+from copy import deepcopy, copy
+import os
+import regex
+from typing import Optional
 
-import parsimonious
+from lark import Lark
+from lark.parsers.lalr_interactive_parser import InteractiveParser
+from lark.parsers.lalr_parser_state import ParserState
+from lark.lexer import Token, LexerState, PatternStr, PatternRE
+from lark.exceptions import UnexpectedCharacters, UnexpectedToken
 
-from lark import Lark, Transformer, v_args
-from parsley import makeGrammar
+
+#########################################################################
+# Fix Lark Speed Issue
+# https://github.com/lark-parser/lark/issues/1142#issuecomment-1863209804
+#########################################################################
+class FastParserState(ParserState):
+    copy_memo = {}
+
+    def __copy__(self):
+        new_value_stack = []
+        for value in self.value_stack:
+            key = f"{id(self)}_{id(value)}"
+            if key not in self.copy_memo:
+                self.copy_memo[key] = deepcopy(value, self.copy_memo)
+            new_value_stack.append(self.copy_memo[key])
+
+        new_instance = type(self)(
+            self.parse_conf,
+            self.lexer, # XXX copy
+            copy(self.state_stack),
+            new_value_stack,
+        )
+
+        self.copy_memo[id(self)] = new_instance
+        return new_instance
 
 
-class TokenIndex:
-    def __init__(self, tokenizer):
-        # map id -> token str including whitespace
+class FastInteractiveParser(InteractiveParser):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.parser_state = FastParserState(
+            self.parser_state.parse_conf,
+            self.parser_state.lexer,
+            self.parser_state.state_stack,
+            self.parser_state.value_stack,
+        )
+
+    def __copy__(self):
+        return type(self)(
+            self.parser,
+            copy(self.parser_state),
+            copy(self.lexer_thread),
+        )
+#########################################################################
+#########################################################################
+
+
+def get_partial_pattern_validator(pattern):
+    """
+    Accepts a pattern object, either lark.lexer.PatternStr or lark.lexer.PatternRE
+    Returns a function which validates a partial string
+
+    e.g. for PatternRE "abc*", returns true for "a", "ab", "abc", "abcccc"
+    """
+    if isinstance(pattern, PatternRE):
+        compiled_pattern = regex.compile(pattern.value)
+        return (
+            lambda seq: compiled_pattern.fullmatch(seq, partial=True) is not None
+        )
+    elif isinstance(pattern, PatternStr):
+        base_str = pattern.value
+        return (
+            lambda seq: base_str.startswith(seq)
+        )
+    else:
+        raise TypeError(f"Invalid pattern type: {type(pattern)}")
+
+
+class InteractivePredictiveLALRParser:
+    """
+    Parser which consumes an EBNF grammar and provides helpers to determine allowable language model tokens
+
+    Interfaces:
+    - step_seq(new_seq): Update the parser with a new sequence to append
+    - is_valid_next_seq(new_seq): Determine whether a candidate sequence is valid
+
+    Core components for terminal level, and sub-terminal level processing:
+    - 1) Lark LALR parser: Applies state transitions, determining set of valid next-terminals
+    - 2) Incremental terminal filter: Eliminates next-terminal candidates if terminal pattern doesn't match
+    """
+
+    def __init__(self, grammar: str, start: str):
+        self.parser = Lark(
+            grammar,
+            regex=True,  # use `regex` not `re`
+            start=start,
+            parser='lalr',
+        )
+        base_interactive_parser = self.parser.parse_interactive()
+        self.interactive_parser = FastInteractiveParser(
+            base_interactive_parser.parser,
+            base_interactive_parser.parser_state,
+            base_interactive_parser.lexer_thread
+        )
+
+        self.partial_seq_validator = {
+            term.name: get_partial_pattern_validator(term.pattern)
+            for term in self.parser.terminals
+        }
+
+        self._ignored_terms = set(self.parser.lexer_conf.ignore)
+
+        # for processing terminals interactively
+        self.last_terminal_pos = 0
+        self.valid_next_terminals = None
+
+        # for calculating `accepts()` efficiently
+        self._accepts_cache = {}
+
+        self.sequence_history = ""
+
+        # initiate
+        self.step_seq("")
+
+    def _accepts(self):
+        if self.sequence_history not in self._accepts_cache:
+            accepted_terminals = self.interactive_parser.accepts()
+            self._accepts_cache[self.sequence_history] = accepted_terminals
+        return self._accepts_cache[self.sequence_history]
+
+    @property
+    def terminal_partial_seq(self):
+        """
+        Return the incomplete subsequence which will eventually comprise a terminal
+        """
+        return self.sequence_history[self.last_terminal_pos:]
+
+    def step_seq(self, new_seq: str):
+        """
+        Append sequence to parser and apply state updates
+        - Append the sequence to the canonical self.sequence_history
+        - Parse the changes
+        - Update the character position of the last complete terminal
+        - Update the set of candidate terminals
+        """
+        self._append_to_sequence(new_seq)
+
+        try:
+            self.interactive_parser.exhaust_lexer()
+        except UnexpectedCharacters as e:
+            self.last_terminal_pos = e.pos_in_stream
+        else:
+            self.last_terminal_pos = len(self.sequence_history)
+
+        self._update_candidate_terminals()
+
+        if not self.valid_next_terminals:
+            raise ValueError(f"Invalid continuation for `{self.sequence_history}` `{new_seq}`")
+
+    def _append_to_sequence(self, new_seq: str):
+        """Set the complete sequences value in the lexer and base"""
+        self.sequence_history += new_seq
+        self.interactive_parser.lexer_thread.state.text = self.sequence_history
+
+    def _update_candidate_terminals(self):
+        """
+        Update the set of candidate terminals
+        - If a new terminal is reached, get the accepted set of terminals from the parser
+        - If the new sequence doesn't comprise a full terminal, filter based on partial pattern match
+        """
+        if not self.terminal_partial_seq:
+            self.valid_next_terminals = self._accepts() | self._ignored_terms
+        else:
+            self.valid_next_terminals = set([
+                term for term in self.valid_next_terminals
+                if self.partial_seq_validator[term](self.terminal_partial_seq)
+            ])
+
+    def is_valid_next_seq(self, new_seq: Optional[str]):
+        """
+        Check if current un-terminalized sequence + new_seq is valid for any terminal
+
+        new_seq can be a string or None representing EOS
+        """
+        if new_seq is None:
+            return "$END" in self.valid_next_terminals
+        for term in self.valid_next_terminals:
+            if term != "$END":
+                if self.partial_seq_validator[term](self.terminal_partial_seq + new_seq):
+                    return True
+        return False
+
+
+class TokenTrie:
+    IS_TOKEN = (None, "is complete token")
+
+    def __init__(self, tokenizer, legal_chars: Optional[set[str]] = None):
+        """
+        Trie structure for efficiently finding tokens which are suffixes of other sequences
+        """
         self.norm_vocab = {}
         for token_id in tokenizer.vocab.values():
-            # TODO: look into difference between tokens, e.g. 28705, 35 are both " 9"
-            # assert norm_token not in self.norm_vocab,
             norm_token = tokenizer.decode([tokenizer.bos_token_id, token_id])[len(tokenizer.bos_token):]
-            self.norm_vocab[norm_token] = token_id
+            if legal_chars is None or all([char in legal_chars for char in norm_token]):
+                self.norm_vocab[norm_token] = token_id
 
-        # get index allowing efficient retrieval of valid tokens given a sequence
-        # given tokens ["art", "artist", "argument", "alice"]
-        # map "a" -> ["ar", "al"]
-        # map "ar" -> ["art", "artist"]
-        # map "art" -> [None, "artist"]  (None indicates match)
-        self.char_map = collections.defaultdict(set)
+        self.token_to_id_set = collections.defaultdict(set)
+        for token_str, token_id in self.norm_vocab.items():
+            self.token_to_id_set[token_str].add(token_id)
+
+        self.trie = {}
         for word in self.norm_vocab:
-            for i in range(1, len(word) + 1):
-                prefix = word[:i]
-                if i < len(word):
-                    self.char_map[prefix].add(word[i])
-                else:
-                    # Add None for complete matches
-                    self.char_map[prefix].add(None)
+            current_dict = self.trie
+            for char in word:
+                if char not in current_dict:
+                    current_dict[char] = {}
+                current_dict = current_dict[char]
+            current_dict[self.IS_TOKEN] = True
 
+    def get_next_level_token_prefixes(self, subprefix: str, _node=None):
+        """
+        Traverse the trie starting from a specified subprefix to identify all child nodes that represent
+        the longest possible strings without omitting any nodes that contain complete tokens.
+        """
+        # if not first level of recursion, and at a branching point or is a token, or return self
+        if _node is not None and (len(_node) > 1 or self.IS_TOKEN in _node):
+            return {subprefix}
 
-    def get_valid_next_charset(self, seq, legal_chars):
-        results = set(self.char_map[seq]) & legal_chars
+        # get the current node if at the first level of recursion
+        if _node is None:
+            _node = self.trie
+            for char in subprefix:
+                if char not in _node:
+                    return set()
+                _node = _node[char]
+
+        # Single child, need to go deeper
+        results = set()
+        for char, next_node in _node.items():
+            if char != self.IS_TOKEN:
+                results |= self.get_next_level_token_prefixes(subprefix + char, _node=next_node)
         return results
 
-    def is_token(self, tok):
-        return tok in self.norm_vocab
+    def is_token(self, seq):
+        return seq in self.norm_vocab
 
 
-class EpsilonNFA:
+class NextTokenValidator:
     """
-    Traverses a Character-Based Epsilon-NFA.
+    Given a grammar and a tokenset, construct a parser and token trie.
 
-    Used by find valid next character sequences.
-
-    self.nfa (dict): A dictionary representing the NFA. It includes:
-        - 'states' (list): A list of states (UUIDsn) inn the NFA.
-        - 'initial_state' (UUID or any hashable ID): The initial state of the NFA.
-        - 'final_states' (list): A list of final or accepting states (UUIDs).
-        - 'alphabets' (list): The set of input symbols (characters).
-        - 'transition_function' (dict): A dictionary representing the state
-          transitions. Each key is a state (UUID), and its value is another
-          dictionary mapping input symbols to lists of next states (UUIDs).
-
-    self.nfa should never be mutated.
+    Interface:
+    - step_seq(new_seq): Append a sequence, update internal states
+    - property valid_token_str_set: The valid set of vocabulary tokens strings which can occur next
     """
-    def __init__(self, nfa):
-        self.nfa = nfa
-
-        # Set of states you may be in
-        self.current_states = set([self.nfa["initial_state"]])
-        self.current_str = ""
-
-        self.legal_chars = set([char for char in self.nfa["alphabets"] if char != "$"])
-
-        self._resolved_epsilon_cache = {}
-
-    def step_seq(self, seq):
-        for char in seq:
-            self.step(char)
-
-    def step(self, char):
-        """
-        Updates the canonical state
-        """
-        next_states = self.next_char_state_map()[char]
-        if not next_states:
-            raise ValueError(f"Illegal transition from '{self.current_str}', no next state for '{char}'")
-        self.current_states = next_states
-        self.current_str += char
-
-    def simulate_step(self, chars, state_set=None):
-        """
-        Return map of chars and their resulting next state-set given a current state-set and new chars
-        """
-        if state_set is None:
-            state_set = self.current_states
-        state_map = self.next_char_state_map(state_set)
-        return {tok: state_map[tok] for tok in chars if tok in state_map}
-
-    def next_char_state_map(self, current_states=None):
-        """
-        Creates a mapping of possible next chars to a set of valid states for each char
-        """
-        if current_states is None:
-            current_states = self.current_states
-
-        char_to_states = collections.defaultdict(set)
-
-        if bool(current_states & set(self.nfa["final_states"])):
-            char_to_states[None] = None
-
-        for state in self._resolve_epsilon_closure(current_states):
-            for char, next_states in self.nfa["transition_function"][state].items():
-                if next_states and char != "$":
-                    char_to_states[char].update(next_states)
-
-        return char_to_states
-
-    def _resolve_epsilon_closure(self, states):
-        closure = set()
-        for state in states:
-            if state in self._resolved_epsilon_cache:
-                new_closures = self._resolved_epsilon_cache[state]
-            else:
-                new_closures = self._get_epsilon_closure(state)
-                self._resolved_epsilon_cache[state] = new_closures
-            closure.update(self._get_epsilon_closure(state))
-        return closure
-
-    def _get_epsilon_closure(self, state, visited=None):
-        if visited is None:
-            visited = set()
-
-        stack = [state]
-        while stack:
-            current_state = stack.pop()
-            if current_state not in visited:
-                visited.add(current_state)
-                stack.extend(self.nfa["transition_function"][current_state].get('$', []))
-
-        return visited
-
-
-class TokenConstraintLogitProcessor:
-    def __init__(self, tokenizer, nfa):
+    def __init__(
+            self,
+            tokenizer,
+            grammar: str,
+            grammar_start: str = "start",
+            num_threads: Optional[int] = None
+    ):
         self.tokenizer = tokenizer
-        self.token_index = TokenIndex(tokenizer)
-        self.nfa = nfa
-        self.prev_token_ids = []
-        self.prev_text = ""
+        self.token_trie = TokenTrie(tokenizer)
+
+        self.parser = InteractivePredictiveLALRParser(
+            grammar=grammar,
+            start=grammar_start
+        )
+
+        # TODO: threading
+        if num_threads is None:
+            self.num_threads = os.cpu_count() // 2
+
+    def step_seq(self, new_seq):
+        self.parser.step_seq(new_seq)
+
+    @property
+    def valid_token_str_set(self):
+        """
+        Generate the set of valid tokens given the current sequence
+
+        1) Push all first level token prefixes to the stack
+        2) for each token in the stack, validate against the parser
+          - if valid, add all children to the stack for later processing
+          - if valid AND a token, add to valid_token_set
+        """
+        valid_token_str_set = set()
+        token_prefix_stack = collections.deque([""])
+        while token_prefix_stack:
+            token_prefix = token_prefix_stack.pop()
+            for child_token_prefix in self.token_trie.get_next_level_token_prefixes(token_prefix):
+                # TODO: Handle EOS token by passing None
+                if self.parser.is_valid_next_seq(child_token_prefix):
+                    token_prefix_stack.append(child_token_prefix)
+                    if self.token_trie.is_token(child_token_prefix):
+                        valid_token_str_set.add(child_token_prefix)
+
+        return valid_token_str_set
+
+    @property
+    def valid_token_id_set(self):
+        """
+        get valid token id based on self.valid_token_str_set
+        note that some token strings correspond to multiple token IDs
+        """
+        return set.union(*[
+            self.token_trie.token_to_id_set[tok]
+            for tok in self.valid_token_str_set
+        ])
+
+
+class GrammarLogitProcessor(NextTokenValidator):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        self.generation_token_ids = []
+        self.generation_text = ""
 
     def __call__(self, token_ids, logits):
 
         # ensure integrity
-        assert token_ids[:len(self.prev_token_ids)] == self.prev_token_ids
-        self.prev_token_ids = token_ids
+        assert token_ids[:len(self.generation_token_ids)] == self.generation_token_ids
+        self.generation_token_ids = token_ids
 
-        # get new text and step NFA forward
-        text = tokenizer.decode(token_ids)
-        new_text = text[len(self.prev_text):]
-        self.prev_text = text
-        self.nfa.step_seq(new_text)
+        # step forward
+        all_text = self.tokenizer.decode(token_ids)
+        new_text = all_text[len(self.generation_text):]
+        self.generation_text = all_text
+        self.step_seq(new_text)
 
-        # get valid new token ids
-        valid_tokens = set(self.get_allowable_next_token_set())
-        valid_token_ids = [
-            self.tokenizer.eos_token_id if t is None else self.token_index.norm_vocab[t]
-            for t in valid_tokens
-        ]
-
-        if not valid_token_ids:
-            raise ValueError("Found no valid tokens, this should never occur.")
-
+        # get valid token IDs and modify logits
+        valid_token_ids = self.valid_token_id_set
         logits = [
             logit_val if tok_id in valid_token_ids else -float("inf")
             for tok_id, logit_val in zip(sorted(self.tokenizer.vocab.values()), logits)
@@ -173,261 +340,199 @@ class TokenConstraintLogitProcessor:
         return logits
 
 
-    def get_allowable_next_token_set(self, current_text="", nfa_next_tok_states=False):
-        """
-        Get set of valid tokens.
-        1) Ask NFA for legal first char
-        3) While legal TokenIndex hasn't been exhausted
-          A) Ask TokenIndex for legal Nth char set
-          B) Ask NFA for
-        """
-        if nfa_next_tok_states is None:
-            return [None]
-        if nfa_next_tok_states == False:
-            nfa_next_tok_states = self.nfa.next_char_state_map()
+def test_generate_json_randomly_via_logit_processor():
+    json_grammar = """
+    ?value: dict
+          | list
+          | string
+          | SIGNED_NUMBER      -> number
+          | "true"             -> true
+          | "false"            -> false
+          | "null"             -> null
 
-        legal_tokens = []
+    list : "[" [value ("," value)*] "]"
 
-        if None in nfa_next_tok_states:
-            legal_tokens.append(None)
-            del nfa_next_tok_states[None]
+    dict : "{" [pair ("," pair)*] "}"
+    pair : string ":" value
 
-        for char, next_states in nfa_next_tok_states.items():
-            # all current sequences are legal per nfa, find legal next token with token index
-            new_seq = current_text + char
-            tokidx_next_chars = self.token_index.get_valid_next_charset(
-                new_seq,
-                self.nfa.legal_chars
-            )
+    string : ESCAPED_STRING
 
-            if self.token_index.is_token(new_seq):
-                legal_tokens.append(new_seq)
-
-            # given legal next chars in token index, get the subset allowed by NFA and recurse
-            legal_tokens += self.get_allowable_next_token_set(
-                new_seq,
-                self.nfa.simulate_step(tokidx_next_chars, next_states)
-            )
-
-        return legal_tokens
-
-
-class LarkPDA:
+    %import common.ESCAPED_STRING
+    %import common.SIGNED_NUMBER
+    %import common.WS
+    #%ignore WS  # we don't ignore whitespace because that makes the json uninteresting
     """
-    Traverses a Lark Parser's PushDown Automata character by character
-    """
-    def __init__(self, lark_grammar: str, start: str = "value"):
-        self.parser = Lark(
-            lark_grammar,
-            start=start,
-            lexer="basic"
+    tokenizer = transformers.AutoTokenizer.from_pretrained("mistralai/Mistral-7B-v0.1")
+
+    logit_processor = GrammarLogitProcessor(
+        tokenizer,
+        json_grammar,
+        grammar_start="value"
+    )
+
+    sample_from_logits = lambda lgts: np.random.choice(len(lgts), p=np.exp(lgts)/np.sum(np.exp(lgts)))
+
+    token_ids = []
+    for _ in range(20):
+        logits = logit_processor(
+            token_ids=token_ids,
+            logits=np.random.uniform(-10, 10, len(tokenizer.vocab))
         )
+        new_token_id = sample_from_logits(logits)
+        token_ids.append(new_token_id)
 
-    def step_seq(self, seq):
-        for char in seq:
-            self.step(char)
-
-    def step(self, char):
-        """
-        Updates the canonical state
-        """
-        next_states = self.next_char_state_map()[char]
-        if not next_states:
-            raise ValueError(f"Illegal transition from '{self.current_str}', no next state for '{char}'")
-        self.current_states = next_states
-        self.current_str += char
-
-    def simulate_step(self, chars, state_set=None):
-        """
-        Return map of chars and their resulting next state-set given a current state-set and new chars
-        """
-        if state_set is None:
-            state_set = self.current_states
-        state_map = self.next_char_state_map(state_set)
-        return {tok: state_map[tok] for tok in chars if tok in state_map}
-
-    def next_char_state_map(self, current_states=None):
-        """
-        Creates a mapping of possible next chars to a set of valid states for each char
-        """
-        if current_states is None:
-            current_states = self.current_states
-
-        char_to_states = collections.defaultdict(set)
-
-        if bool(current_states & set(self.nfa["final_states"])):
-            char_to_states[None] = None
-
-        for state in self._resolve_epsilon_closure(current_states):
-            for char, next_states in self.nfa["transition_function"][state].items():
-                if next_states and char != "$":
-                    char_to_states[char].update(next_states)
-
-        return char_to_states
-
-
-def lark_to_pushdown_automata_spec(lark_grammar: str, start: str = "value") -> dict:
-    parser = Lark(lark_grammar, start=start, lexer="basic")
     import pdb;pdb.set_trace()
 
 
-def handle_regex(pattern):
+def test_next_token_validator_simple():
+    hello_grammar = """
+    ?value: "hello" | "world"
     """
-    Handle a regex pattern and convert it into a state representation for PDA.
+    tokenizer = transformers.AutoTokenizer.from_pretrained("mistralai/Mistral-7B-v0.1")
+    ntv = NextTokenValidator(tokenizer, hello_grammar, "value")
 
-    Args:
-    pattern (str): The pattern of the regular expression.
+    assert ntv.valid_token_str_set == {'wo', 'hell', 'h', 'he', 'hel', 'world', 'wor', 'w', 'hello'}
+    assert ntv.valid_token_id_set == {265, 809, 107, 2805, 21558, 28727, 13436, 22493, 9471}
 
-    Returns:
-    str: A unique representation of the regex for the PDA.
+
+def test_token_trie_sanity_hf_tokenizer():
+    """Ensure token trie produces the same number of N 3 letter tokens"""
+    tokenizer = transformers.AutoTokenizer.from_pretrained("mistralai/Mistral-7B-v0.1")
+    toktrie = TokenTrie(tokenizer)
+
+    all_prefixes = toktrie.get_next_level_token_prefixes("")
+
+    # every token should be composable from a single unique char, so they will all be len of 1
+    assert all([len(p) == 1 for p in all_prefixes])
+
+    # every token should have one of these prefixes as a start character
+    assert all([
+        t[0] in all_prefixes
+        for t in toktrie.norm_vocab
+    ])
+
+    # construct the set of next level prefixes
+    all_subprefixes = set()
+    for pfx in all_prefixes:
+        all_subprefixes |= toktrie.get_next_level_token_prefixes(pfx)
+
+    # these should have varying length because some tokens don't have level-2 prefixes
+    assert len(set([len(spfx) for spfx in all_subprefixes])) > 1
+
+
+def test_simple_sequence(parser):
+    for token in ['{', '"', 'k', 'ey', '":', '"val', 'ue', '"']:
+        print("full:", parser.sequence_history)
+        print("adding", token)
+        parser.step_seq(token)
+        print("partial:", parser.terminal_partial_seq)
+        print("valid terms:", parser.valid_next_terminals)
+
+
+def test_valid_next_tokens():
+    json_grammar = """
+    ?value: dict
+          | list
+          | string
+          | SIGNED_NUMBER      -> number
+          | "true"             -> true
+          | "false"            -> false
+          | "null"             -> null
+
+    list : "[" [value ("," value)*] "]"
+
+    dict : "{" [pair ("," pair)*] "}"
+    pair : string ":" value
+
+    string : ESCAPED_STRING
+
+    %import common.ESCAPED_STRING
+    %import common.SIGNED_NUMBER
+    %import common.WS
+    %ignore WS
     """
-    # Convert the regex pattern into a suitable representation for the PDA.
-    # This is a placeholder for the actual conversion logic.
-    return f"REGEX({pattern})"
+
+    parser = InteractivePredictiveLALRParser(json_grammar, 'value')
+    # random complicated json file courtesy of https://github.com/simdjson/simdjson/issues/1316#issue-748663718
+    complex_json_file = '{"$schema": "http://json-schema.org/draft-04/schema#", "additionalProperties": false, "properties": {"nc:Vehicle": {"description": "A conveyance designed to carry an operator, passengers and/or cargo, over land.", "oneOf": [{"$ref": "#/definitions/nc:VehicleType"}, {"type": "array", "items": {"$ref": "#/definitions/nc:VehicleType"}}]}, "nc:VehicleAxleQuantity": {"description": "A count of common axles of rotation of one or more wheels of a vehicle, whether power driven or freely rotating.", "oneOf": [{"$ref": "#/definitions/niem-xs:nonNegativeInteger"}, {"type": "array", "items": {"$ref": "#/definitions/niem-xs:nonNegativeInteger"}}]}, "nc:VehicleMSRPAmount": {"description": "A manufacturer\'s suggested retail price of a vehicle; a price at which a manufacturer recommends a vehicle be sold.", "oneOf": [{"$ref": "#/definitions/nc:AmountType"}, {"type": "array", "items": {"$ref": "#/definitions/nc:AmountType"}}]}, "nc:Amount": {"description": "An amount of money.", "oneOf": [{"$ref": "#/definitions/niem-xs:decimal"}, {"type": "array", "items": {"$ref": "#/definitions/niem-xs:decimal"}}]}, "nc:Currency": {"description": "A data concept for a unit of money or exchange.", "oneOf": [{"anyOf": [{"$ref": "#/properties/nc:CurrencyCode"}]}, {"type": "array", "items": {"anyOf": [{"$ref": "#/properties/nc:CurrencyCode"}]}}]}, "nc:CurrencyCode": {"description": "A unit of money or exchange.", "oneOf": [{"$ref": "#/definitions/iso_4217:CurrencyCodeType"}, {"type": "array", "items": {"$ref": "#/definitions/iso_4217:CurrencyCodeType"}}]}, "nc:VehicleIdentification": {"description": "A unique identification for a specific vehicle.", "oneOf": [{"$ref": "#/definitions/nc:IdentificationType"}, {"type": "array", "items": {"$ref": "#/definitions/nc:IdentificationType"}}]}, "nc:IdentificationID": {"description": "An identifier.", "oneOf": [{"$ref": "#/definitions/niem-xs:string"}, {"type": "array", "items": {"$ref": "#/definitions/niem-xs:string"}}]}}, "definitions": {"nc:VehicleType": {"description": "A data type for a conveyance designed to carry an operator, passengers and/or cargo, over land.", "allOf": [{"$ref": "#/definitions/nc:ConveyanceType"}, {"type": "object", "properties": {"nc:VehicleAxleQuantity": {"$ref": "#/properties/nc:VehicleAxleQuantity"}, "nc:VehicleIdentification": {"$ref": "#/properties/nc:VehicleIdentification"}, "nc:VehicleMSRPAmount": {"$ref": "#/properties/nc:VehicleMSRPAmount"}}}]}, "nc:ConveyanceType": {"description": "A data type for a means of transport from place to place.", "allOf": [{"$ref": "#/definitions/_base"}, {"$ref": "#/definitions/nc:ItemType"}, {"type": "object", "properties": {}}]}, "nc:ItemType": {"description": "A data type for an article or thing.", "allOf": [{"$ref": "#/definitions/_base"}, {"type": "object", "properties": {}}]}, "nc:AmountType": {"description": "A data type for an amount of money.", "type": "object", "properties": {"nc:Amount": {"$ref": "#/properties/nc:Amount"}, "nc:Currency": {"$ref": "#/properties/nc:Currency"}}}, "iso_4217:CurrencyCodeType": {"description": "A data type for a currency that qualifies a monetary amount.", "oneOf": [{"$ref": "#/definitions/iso_4217:CurrencyCodeSimpleType"}, {"type": "object", "properties": {"rdf:value": {"$ref": "#/definitions/iso_4217:CurrencyCodeSimpleType"}}}]}, "iso_4217:CurrencyCodeSimpleType": {"type": "string", "description": "A data type for a currency that qualifies a monetary amount.", "oneOf": [{"enum": ["EUR"], "description": "Euro"}, {"enum": ["GBP"], "description": "Pound Sterling"}, {"enum": ["USD"], "description": "US Dollar"}]}, "nc:IdentificationType": {"description": "A data type for a representation of an identity.", "type": "object", "properties": {"nc:IdentificationID": {"$ref": "#/properties/nc:IdentificationID"}}}, "niem-xs:decimal": {"description": "A data type for arbitrary precision decimal numbers.", "type": "number"}, "niem-xs:nonNegativeInteger": {"description": "A data type for an integer with a minimum value of 0.", "type": "number"}, "niem-xs:string": {"description": "A data type for character strings in XML.", "type": "string"}, "_base": {"type": "object", "patternProperties": {"^ism:.*": {"type": "string"}, "^ntk:.*": {"type": "string"}}, "properties": {"@id": {"format": "uriref"}, "@base": {"format": "uriref"}}}}}'
+
+    test_chars_per_iter = 1000
+    unicode_chars = [chr(i) for i in range(test_chars_per_iter)]
+
+    import time
+    start = time.time()
+    for char in complex_json_file:
+        parser.step_seq(char)
+        for ch in unicode_chars:
+            parser.is_valid_next_seq(ch)
+
+    print("took",
+          (time.time() - start) /  (len(complex_json_file)),
+          "seconds per step with",
+          test_chars_per_iter, "characters in vocabulary")
 
 
-from parsimonious.nodes import NodeVisitor
-from parsimonious.expressions import Literal, Sequence, Regex
 
-def convert_grammar_to_pda(grammar):
-    pda = {
-        'states': set(),
-        'initial_state': 'q0',
-        'final_states': {'q_accept'},
-        'stack_alphabet': set(),
-        'transition_function': collections.defaultdict(lambda: collections.defaultdict(list))
-    }
+def profile_predictor():
+    import pstats
+    from io import StringIO
+    import cProfile
+    hello_grammar = """
+    ?value: "hello" | "world"
+    """
+    tokenizer = transformers.AutoTokenizer.from_pretrained("mistralai/Mistral-7B-v0.1")
+    ntv = NextTokenValidator(tokenizer, hello_grammar, "value")
 
-    def process_expression(state, expr, is_terminal=False):
-        if isinstance(expr, Literal):
-            pda['stack_alphabet'].add(expr.literal)
-            next_state = 'q_accept' if is_terminal else 'q' + str(len(pda['states']) + 1)
-            pda['transition_function'][state][('epsilon', expr.literal)].append((next_state, []))
-            if not is_terminal:
-                pda['states'].add(next_state)
-        elif isinstance(expr, Sequence):
-            current_state = state
-            for i, member in enumerate(expr.members):
-                is_last_member = i == len(expr.members) - 1
-                process_expression(current_state, member, is_terminal=is_last_member)
-                if not is_last_member:
-                    current_state = 'q' + str(len(pda['states']))
-        elif isinstance(expr, Regex):
-            raise NotImplementedError("Regex handling not implemented")
+    profile = cProfile.Profile()
+    profile.enable()
+    #####
 
-    for non_terminal, production in grammar.items():
-        pda['states'].add(non_terminal)
-        pda['stack_alphabet'].add(non_terminal)
-        process_expression(non_terminal, production)
+    valid_toks = ntv.valid_token_str_set
+    ntv.step_seq("h")
+    valid_toks = ntv.valid_token_str_set
+    ntv.step_seq("e")
+    valid_toks = ntv.valid_token_str_set
+    ntv.step_seq("l")
+    valid_toks = ntv.valid_token_str_set
+    ntv.step_seq("l")
+    valid_toks = ntv.valid_token_str_set
+    ntv.step_seq("o")
 
-    return dict(pda)
+    #####
+    profile.disable()
+
+    # Sorting the statistics by cumulative time
+    s = StringIO()
+    sortby = 'cumulative'
+    ps = pstats.Stats(profile, stream=s).sort_stats(sortby)
+    ps.print_stats()
+    print(s.getvalue())
 
 
-# Test Case
-def test_convert_simple_grammar_to_pda():
-    grammar = Grammar("""
-    expression = term "+" term
-    term       = "number"
-    """)
 
-    expected_pda = {
-        'states': {'expression', 'term', 'q0', 'q_accept'},
-        'initial_state': 'q0',
-        'final_states': {'q_accept'},
-        'stack_alphabet': {'expression', 'term', '+', 'number'},
-        'transition_function': {
-            'q0': {('epsilon', 'epsilon'): [('expression', ['term', '+', 'term'])]},
-            'expression': {('epsilon', 'term'): [('term', [])]},
-            'term': {('epsilon', 'number'): [('q_accept', [])]}
-        }
-    }
-
-    actual_pda = convert_grammar_to_pda(grammar)
-    import pprint
-    pprint.pprint(actual_pda)
-    assert actual_pda == expected_pda, "PDA configuration does not match expected output"
+def main():
+    test_generate_json_randomly_via_logit_processor()
 
 
 if __name__ == "__main__":
     import transformers
     import numpy as np
 
-    test_convert_simple_grammar_to_pda()
-    import pdb;pdb.set_trace()
+    profile = False
+    if profile:
+        import cProfile
+        import pstats
+        from io import StringIO
+        profile = cProfile.Profile()
+        profile.enable()
+        main()
+        profile.disable()
 
-
-    grammar = r"""
-    expr        = (entry / emptyline)*
-    entry       = section pair*
-
-    section     = lpar word rpar ws
-    pair        = key equal value ws?
-
-    key         = word+
-    value       = (word / quoted)+
-    word        = ~r"[-\w]+"
-    quoted      = ~'"[^\"]+"'
-    equal       = ws? "=" ws?
-    lpar        = "["
-    rpar        = "]"
-    ws          = ~"\s*"
-    emptyline   = ws+
-    """
-
-    pda_config = convert_grammar_to_pda(Grammar(grammar))
-    import pdb;pdb.set_trace()
-
-    try_parsley()
-
-    json_grammar = r"""
-        ?start: value
-
-        ?value: object
-              | array
-              | string
-              | SIGNED_NUMBER      -> number
-              | "true"             -> true
-              | "false"            -> false
-              | "null"             -> null
-
-        array  : "[" [value ("," value)*] "]"
-        object : "{" [pair ("," pair)*] "}"
-        pair   : string ":" value
-
-        string : ESCAPED_STRING
-
-        %import common.ESCAPED_STRING
-        %import common.SIGNED_NUMBER
-        %import common.WS
-
-        %ignore WS
-    """
-    #lark_to_pushdown_automata_spec(json_grammar)
-
-    tokenizer = transformers.AutoTokenizer.from_pretrained("mistralai/Mistral-7B-v0.1")
-
-    sample_from_logits = lambda lgts: np.random.choice(len(lgts), p=np.exp(lgts)/np.sum(np.exp(lgts)))
-
-    for i in range(4):
-
-        logit_processor = TokenConstraintLogitProcessor(
-            tokenizer=tokenizer,
-            nfa=EpsilonNFA(nfa=regex_to_nfa.regex_to_nfa(
-                r"(large )?(language )((models )+(inference engines ))(are )((useful)+((very )*complex))."
-            )),
-        )
-
-        token_ids = []
-        while True:
-            logits = logit_processor(
-                token_ids=token_ids,
-                logits=np.random.uniform(-10, 10, len(tokenizer.vocab))
-            )
-            new_token_id = sample_from_logits(logits)
-            token_ids.append(new_token_id)
-            if new_token_id == tokenizer.eos_token_id:
-                break
-        print(f"run #{i}")
-        print("\ttokenid", token_ids)
-        print("\ttokens:", [tokenizer.decode(tok_id, ) for tok_id in token_ids])
-        print("\tresult:", tokenizer.decode(token_ids, skip_special_tokens=False))
+        # Sorting the statistics by cumulative time
+        s = StringIO()
+        sortby = 'cumulative'
+        ps = pstats.Stats(profile, stream=s).sort_stats(sortby)
+        ps.print_stats()
+        print(s.getvalue())
+    else:
+        main()
